refactor: discount 패키지명을 benefit으로 수정, BenefitHolder 클래스를 해당 패키지로 이동
