+++ src/main/java/christmas/model/benefit/GiftBenefit.java
@@ -1,4 +1,4 @@
-package christmas.model.discount;
+package christmas.model.benefit;
 
 import christmas.model.Gift;
