+++ src/test/java/christmas/model/benefit/GiftBenefitTest.java
@@ -1,0 +1,6 @@
+package christmas.model.benefit;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class GiftBenefitTest {
+}
