--- src/test/java/christmas/model/discount/GiftBenefitTest.java
@@ -1,12 +1,0 @@
-package christmas.model.discount;
-
-import christmas.model.Gift;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.when;
-
-class GiftBenefitTest {
-}
